+++ dotNES/Mappers/Mapper180.cs
@@ -12,7 +12,7 @@ namespace dotNES.Mappers
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public override byte ReadByte(uint addr)
+        public override uint ReadByte(uint addr)
         {
             if (addr < 0x8000) return _RAM[addr - 0x6000];
             if (addr < 0xC000) return _PRGROM[addr - 0x8000];
